refactor(agents): inline stream runtime preparation
