+++ src/agents/embedded-agent-runner/run/attempt-execution-phase.ts
@@ -1,12 +1,24 @@
 /** Prepares the guarded stream runtime before prompt execution and settlement. */
 import {
+  bindOwnedSessionTranscriptWrites,
+  withOwnedSessionTranscriptWrites,
+} from "../../../config/sessions/transcript-write-context.js";
+import {
   mergeAgentRunAttemptTerminal,
   projectAgentRunAttemptTerminal,
   type AgentRunAttemptTerminal,
 } from "../../agent-run-terminal-outcome.js";
+import { log } from "../logger.js";
+import type { EmbeddedAgentQueueHandle } from "../runs.js";
+import { flushPendingToolResultsAfterIdle } from "../wait-for-idle-before-flush.js";
+import { abortable as abortableWithSignal } from "./abortable.js";
 import type { EmbeddedAttemptExecutionPhaseInput } from "./attempt-execution-types.js";
+import { createEmbeddedAttemptRunAbort } from "./attempt-finalize.js";
+import { prepareEmbeddedAttemptHistory } from "./attempt-history.js";
 import { runEmbeddedAttemptSettledPhase } from "./attempt-settle.js";
-import { prepareEmbeddedAttemptStreamRuntime } from "./attempt-stream-runtime-prepare.js";
+import { prepareEmbeddedAttemptStream } from "./attempt-stream-prepare.js";
+import { installEmbeddedAttemptStreamGuards } from "./attempt-stream.js";
+import { prepareEmbeddedAttemptTimeout } from "./attempt-timeout-prepare.js";
 import type { EmbeddedRunAttemptResult } from "./types.js";
 
 export async function runEmbeddedAttemptExecutionPhase(
@@ -49,33 +61,40 @@ export async function runEmbeddedAttemptExecutionPhase(
     state.terminal = mergeAgentRunAttemptTerminal(state.terminal, incoming);
   };
 
-  const preparedStreamRuntime = await prepareEmbeddedAttemptStreamRuntime({
+  const idleTimeoutTriggerRef: { current?: (error: Error) => void } = {};
+  const { cacheObservabilityEnabled, promptCacheTools } = installEmbeddedAttemptStreamGuards({
     attempt,
-    activeSession,
+    session: activeSession,
     sessionManager,
-    ownedTranscriptWriteContext: input.sessionLock.ownedTranscriptWriteContext,
-    runAbortController: input.runAbortController,
-    externalAbortController: input.externalAbortController,
-    abortActiveSession,
-    abortState: input.abortState,
-    trackPromptSettlePromise,
-    compactionTimeoutMs: input.sessionLock.compactionTimeoutMs,
-    guards: {
-      sessionAgentId: input.setup.sessionAgentId,
-      cacheTrace,
-      allCustomTools,
-      systemPromptText: sessionRuntimeState.systemPromptText,
-      transcriptPolicy,
-      isOpenAIResponsesApi,
-      replayAllowedToolNames,
-      liveAllowedToolNames,
-      clientToolLoopDetection,
-      anthropicPayloadLogger,
-      effectiveAgentTransport,
-      providerTextTransforms,
-      runTrace: input.diagnostics.runTrace,
+    sessionAgentId: input.setup.sessionAgentId,
+    cacheTrace,
+    allCustomTools,
+    systemPromptText: sessionRuntimeState.systemPromptText,
+    transcriptPolicy,
+    isOpenAIResponsesApi,
+    replayAllowedToolNames,
+    liveAllowedToolNames,
+    clientToolLoopDetection,
+    anthropicPayloadLogger,
+    effectiveAgentTransport,
+    providerTextTransforms,
+    runTrace: input.diagnostics.runTrace,
+    isYieldDetected: () => input.lifecycle.readYieldState().yieldDetected,
+    onRejectedThinkingReplayRepaired: () => {
+      repairedRejectedThinkingReplay = true;
     },
-    history: {
+    onIdleTimeout: (error) => idleTimeoutTriggerRef.current?.(error),
+    abortSignal: input.runAbortController.signal,
+  });
+  input.setup.prepStages.mark("stream-setup");
+  input.setup.emitPrepStageSummary("stream-ready");
+
+  let preparedHistory: Awaited<ReturnType<typeof prepareEmbeddedAttemptHistory>>;
+  try {
+    preparedHistory = await prepareEmbeddedAttemptHistory({
+      attempt,
+      activeSession,
+      sessionManager,
       ...(input.activeContextEngine ? { activeContextEngine: input.activeContextEngine } : {}),
       cacheTrace,
       capabilityToolNames,
@@ -90,48 +109,124 @@ export async function runEmbeddedAttemptExecutionPhase(
       systemPromptText: sessionRuntimeState.systemPromptText,
       transcriptPolicy,
       setActiveSessionSystemPrompt,
-    },
-    stream: {
-      runtimeChannel,
-      hookRunner,
-      hookAgentId,
-      diagnosticTrace: input.diagnostics.diagnosticTrace,
-      clientToolCallSlots,
-      toolSearchTargetTranscriptProjections,
-      isReplaySafeTool: (tool) => replaySafeTools.has(tool as never),
-      hasDeliveredSourceReply,
-      markSourceReplyDelivered,
-      sandboxSessionKey: input.setup.sandboxSessionKey,
-      builtinToolNames,
-      replaySafeToolNames,
-    },
-    lifecycle: {
-      isYieldDetected: () => input.lifecycle.readYieldState().yieldDetected,
-      markRejectedThinkingReplayRepaired: () => {
-        repairedRejectedThinkingReplay = true;
-      },
-      markStreamReady: () => {
-        input.setup.prepStages.mark("stream-setup");
-        input.setup.emitPrepStageSummary("stream-ready");
-      },
-      markIdleTimedOut: () => mergeTerminal({ kind: "timeout", phase: "prompt", source: "idle" }),
-      markExternalAbort: () => mergeTerminal({ kind: "aborted", source: "external" }),
-      markTimedOutDuringCompaction: () =>
-        mergeTerminal({ kind: "timeout", phase: "compaction", source: "observation" }),
-      markTimedOutByRunBudget: () =>
-        mergeTerminal({ kind: "timeout", phase: "prompt", source: "run_budget" }),
-      readRunState: () => {
-        const terminal = projectAgentRunAttemptTerminal(state.terminal);
-        return {
-          aborted: terminal.aborted,
-          promptError: terminal.promptError,
-          timedOut: terminal.timedOut,
-          yieldDetected: input.lifecycle.readYieldState().yieldDetected,
-        };
-      },
-      setToolSearchCatalogExecutor: input.lifecycle.setToolSearchCatalogExecutor,
-    },
+    });
+  } catch (error) {
+    await flushPendingToolResultsAfterIdle({
+      agent: activeSession.agent,
+      sessionManager,
+      // An already-aborted setup must dispose immediately without orphaning tool calls.
+      ...(attempt.abortSignal?.aborted ? { timeoutMs: 0 } : {}),
+    });
+    activeSession.dispose();
+    throw error;
+  }
+
+  const isProbeSession = attempt.sessionId?.startsWith("probe-") ?? false;
+  const queueHandleRef: { current?: EmbeddedAgentQueueHandle } = {};
+  const abortRun = createEmbeddedAttemptRunAbort({
+    abortActiveSession,
+    activeSession,
+    attempt,
+    getQueueHandle: () => queueHandleRef.current,
+    isProbeSession,
+    log,
+    runAbortController: input.runAbortController,
+    state: input.abortState,
   });
+  input.externalAbortController.setRunAbort(abortRun);
+  idleTimeoutTriggerRef.current = (error) => {
+    mergeTerminal({ kind: "timeout", phase: "prompt", source: "idle" });
+    abortRun(true, error);
+  };
+  const abortable = <T>(promise: Promise<T>): Promise<T> =>
+    abortableWithSignal(input.runAbortController.signal, promise);
+  const promptActiveSession = (
+    prompt: string,
+    options?: Parameters<typeof activeSession.prompt>[1],
+  ): Promise<void> =>
+    withOwnedSessionTranscriptWrites(input.sessionLock.ownedTranscriptWriteContext, async () => {
+      // Prompting starts its own agent loop; reject before creating a loop that
+      // an already-aborted attempt can no longer cancel.
+      if (input.runAbortController.signal.aborted) {
+        return abortable(Promise.resolve());
+      }
+      return abortable(trackPromptSettlePromise(activeSession.prompt(prompt, options)));
+    });
+  const onBlockReply = attempt.onBlockReply
+    ? bindOwnedSessionTranscriptWrites(
+        input.sessionLock.ownedTranscriptWriteContext,
+        attempt.onBlockReply,
+      )
+    : undefined;
+  const onBlockReplyFlush = attempt.onBlockReplyFlush
+    ? bindOwnedSessionTranscriptWrites(
+        input.sessionLock.ownedTranscriptWriteContext,
+        attempt.onBlockReplyFlush,
+      )
+    : undefined;
+  const preparedStream = prepareEmbeddedAttemptStream({
+    attempt,
+    activeSession,
+    runAbortController: input.runAbortController,
+    abortRun,
+    markExternalAbort: () => mergeTerminal({ kind: "aborted", source: "external" }),
+    getRunState: () => {
+      const terminal = projectAgentRunAttemptTerminal(state.terminal);
+      return {
+        aborted: terminal.aborted,
+        promptError: terminal.promptError,
+        timedOut: terminal.timedOut,
+        yieldDetected: input.lifecycle.readYieldState().yieldDetected,
+      };
+    },
+    onBlockReply,
+    onBlockReplyFlush,
+    runtimeChannel,
+    hookRunner,
+    hookAgentId,
+    diagnosticTrace: input.diagnostics.diagnosticTrace,
+    clientToolCallSlots,
+    toolSearchTargetTranscriptProjections,
+    isReplaySafeTool: (tool) => replaySafeTools.has(tool as never),
+    hasDeliveredSourceReply,
+    markSourceReplyDelivered,
+    sandboxSessionKey: input.setup.sandboxSessionKey,
+    builtinToolNames,
+    replaySafeToolNames,
+  });
+  input.lifecycle.setToolSearchCatalogExecutor(preparedStream.toolSearchCatalogExecutor);
+  input.externalAbortController.setCompactionState({
+    isPendingOrRetrying: preparedStream.subscription.isCompacting,
+    isInFlight: () => activeSession.isCompacting,
+  });
+  queueHandleRef.current = preparedStream.queueHandle;
+
+  const attemptTimeout = prepareEmbeddedAttemptTimeout({
+    attempt,
+    activeSession,
+    compactionState: preparedStream.subscription,
+    compactionTimeoutMs: input.sessionLock.compactionTimeoutMs,
+    isProbeSession,
+    abortRun,
+    markTimedOutDuringCompaction: () =>
+      mergeTerminal({ kind: "timeout", phase: "compaction", source: "observation" }),
+    markTimedOutByRunBudget: () =>
+      mergeTerminal({ kind: "timeout", phase: "prompt", source: "run_budget" }),
+  });
+
+  const preparedStreamRuntime = {
+    abortable,
+    cache: {
+      observabilityEnabled: cacheObservabilityEnabled,
+      promptTools: promptCacheTools,
+    },
+    history: preparedHistory,
+    isProbeSession,
+    onBlockReplyFlush,
+    promptActiveSession,
+    stream: preparedStream,
+    timeout: attemptTimeout,
+  };
   return await runEmbeddedAttemptSettledPhase({
     ...input,
     preparedStreamRuntime,
